backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
langchain>=0.1.0
langchain-experimental>=0.0.50
langchain-groq>=0.0.1
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
tabulate>=0.9.0
chardet>=5.0.0